Update blockly.js
Added a field "tags" to hand over a tag to pushover-service. With this it is possible to revoke a message with prio 2 / alert. Helpful if you have an active alert and your system notice, that this alert isn't valid any more. 

An alarm have to be withdrawn manually with curl.